Reimplement stackinfo handling
Previous implementation was not very usable. This new implementation
allows the stackinfo to be re-initialized at the first layer, whatever
protocol the first layer may be.

The idea is to pass NULL to the handler function and check for NULL in
each handler function. If the user data being passed is NULL, it means
that the handler is handling the first layer. It should then call
stackinfo_new() to re-initialize the stattic buffer, and pass the
pointer to this buffer (returned by the function) to the next handler.