Add basic packet handler
This is a handler that handles each packet before the first protocol
handler is called. This is useful to setup the stackinfo structure so
that the protocol handlers do not need the if/else construction. It is
also useful to perform tasks that is unrelated to any of the protocols
being handled, like printing generic packet information.